fix: update springboot3 server mongo configuration

+++ src/main/java/net/petstore/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package net.petstore.service;
 
+import lombok.extern.slf4j.Slf4j;
 import net.petstore.model.User;
 import net.petstore.repository.UserRepository;
 import org.modelmapper.ModelMapper;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
+@Slf4j
 public class UserServiceImpl implements  UserService{
 
     @Autowired
@@ -19,50 +21,61 @@ public class UserServiceImpl implements  UserService{
 
 
     public void createUser(User body) {
-        //  save the user to the database
-        net.petstore.domain.User userDTO = modelMapper.map(body,  net.petstore.domain.User.class);
-        userRepository.save(userDTO);
-
+        // log request
+        log.info("createUser: {}", body);
+        // convert user model to user entity
+        net.petstore.domain.User user = modelMapper.map(body, net.petstore.domain.User.class);
+        // save user entity
+        userRepository.save(user);
 
     }
 
 
     public void deleteUser( String username) {
-        //  find the user by username and delete it
-
+        log.info("deleteUser: {}", username);
         net.petstore.domain.User user = userRepository.findByUsername(username);
-        userRepository.delete(user);
+        if (user != null) {
+            userRepository.delete(user);
+        }
 
     }
 
     public User getUserByName( String username) {
+        log.info("getUserByName: {}", username);
+
+
         net.petstore.domain.User byUsername = userRepository.findByUsername(username);
+        // convert user entity to user model
         return modelMapper.map(byUsername, User.class);
-        
+
     }
 
     public void updateUser( String username,User body) {
-        //  find the user by username and update it
+        log.info("updateUser: {}", username);
+        // find user by username
         net.petstore.domain.User user = userRepository.findByUsername(username);
-        user.setFirstName(body.getFirstName());
-        user.setLastName(body.getLastName());
-        user.setEmail(body.getEmail());
-        user.setPassword(body.getPassword());
-        userRepository.save(user);
-
+        if (user != null) {
+            // update user entity
+            user.setFirstName(body.getFirstName());
+            user.setLastName(body.getLastName());
+            user.setEmail(body.getEmail());
+            user.setPassword(body.getPassword());
+            userRepository.save(user);
+        }
     }
 
     public void createUsersWithArrayInput(List<User> body) {
-      // use a stream to save all the users in the list
-        body.forEach(user -> {
-            net.petstore.domain.User userDTO = modelMapper.map(user,  net.petstore.domain.User.class);
-            userRepository.save(userDTO);
-        });
-
+        // save all users
+        for (User user: body) {
+            createUser(user);
+        }
     }
 
     public void createUsersWithListInput(List<User> body) {
-        createUsersWithArrayInput(body);
+        // save all users
+        for (User user: body) {
+            createUser(user);
+        }
     }
